+++ amazonorders/cli.py
@@ -319,8 +319,9 @@ def _authenticate(amazon_session: AmazonSession,
     except AmazonOrdersAuthError as e:
         if retries < 1:
             if amazon_session.username:
+                click.secho(str(f"{e}\n"), fg="red")
                 click.echo(
-                    f"Info: Authenticating '{amazon_session.username}' failed, retrying ...\n")
+                    f"Info: Authenticating '{amazon_session.username}' failed. If using 2FA, wait a minute before retrying with a new OTP.\n")
 
             amazon_session.password = None