+++ tests/integration/test_integration_generic.py
@@ -2,108 +2,35 @@ __copyright__ = "Copyright (c) 2024-2025 Alex Laird"
 __license__ = "MIT"
 
 import datetime
-import json
 import os
-import shutil
-import unittest
 
-from amazonorders.conf import AmazonOrdersConfig
-from amazonorders.exception import AmazonOrdersNotFoundError, AmazonOrdersAuthError
-from amazonorders.orders import AmazonOrders
-from amazonorders.session import AmazonSession
-from amazonorders.transactions import AmazonTransactions
+from amazonorders.exception import AmazonOrdersNotFoundError
 from tests.integrationtestcase import IntegrationTestCase
 
 
 class TestIntegrationGeneric(IntegrationTestCase):
     """
     These integration tests run generically against any Amazon account. The only requirement is that the
-    account in question has at least one order in the year ``AMAZON_INTEGRATION_TEST_YEAR`` (defaults to the
-    current year). The only assertions done on the fields populated are ``isNotNone``.
+    account in question have at least one Order and one Transaction in the year ``AMAZON_INTEGRATION_TEST_YEAR``
+    (defaults to the current year). The only assertions done on the fields populated are to test if values are
+    set generically, not asserting on specific values.
     """
 
     @classmethod
     def setUpClass(cls):
         super().setUpClass()
 
-        if os.environ.get("AMAZON_INTEGRATION_TEST_YEAR"):
-            cls.year = os.environ.get("AMAZON_INTEGRATION_TEST_YEAR")
-        else:
+        cls.year = os.environ.get("AMAZON_INTEGRATION_TEST_YEAR")
+        if not cls.year:
             cls.year = datetime.date.today().year
-        if os.environ.get("AMAZON_START_INDEX"):
-            cls.start_index = os.environ.get("AMAZON_START_INDEX")
-        else:
-            cls.start_index = None
-        if os.environ.get("AMAZON_TRANSACTIONS_DAYS"):
-            cls.transactions_days = os.environ.get("AMAZON_TRANSACTIONS_DAYS")
-        else:
+        cls.start_index = os.environ.get("AMAZON_START_INDEX")
+        cls.transactions_days = os.environ.get("AMAZON_TRANSACTIONS_DAYS")
+        if not cls.transactions_days:
             cls.transactions_days = 90
         if os.environ.get("AMAZON_FULL_DETAILS_LOOP_COUNT"):
             cls.full_details_loop_count = int(os.environ.get("AMAZON_FULL_DETAILS_LOOP_COUNT"))
         else:
             cls.full_details_loop_count = 1
-
-    @unittest.skip("Even though the form submissions suggest tha Amazon should not re-prompt for OTP, it often does. "
-                   "Further investigation needs to be done, perhaps in to deviceId, for this test to pass.")
-    def test_logout_login_no_otp_reprompt(self):
-        # GIVEN
-        with open(self.test_config.cookie_jar_path, "r") as f:
-            logged_in_persisted_cookies = f.read()
-
-        # WHEN
-        self.amazon_session.logout()
-
-        # THEN
-        with open(self.test_config.cookie_jar_path, "r") as f:
-            logged_out_persisted_cookies = f.read()
-        self.assertFalse(self.amazon_session.is_authenticated)
-        self.assertNotEqual(logged_in_persisted_cookies, logged_out_persisted_cookies)
-
-        # GIVEN
-        self.amazon_session.otp_secret_key = None
-
-        # WHEN
-        # If the test is not automated (ie. prompts for OTP here), consider that a failure
-        self.amazon_session.login()
-
-        # THEN
-        self.assertTrue(self.amazon_session.is_authenticated)
-
-    def test_persisted_session_expired(self):
-        # GIVEN
-        test_config_dir = os.path.join(os.path.abspath(os.path.dirname(__file__)), "..",
-                                       ".integration-config-auth-test")
-        if os.path.exists(test_config_dir):
-            shutil.rmtree(test_config_dir)
-        test_output_dir = os.path.join(test_config_dir, "output")
-        test_cookie_jar_path = os.path.join(test_config_dir, "cookies.json")
-        test_config = AmazonOrdersConfig(data={
-            "output_dir": test_output_dir,
-            "cookie_jar_path": test_cookie_jar_path
-        })
-        amazon_session = AmazonSession(debug=os.environ.get("DEBUG", "False") == "True",
-                                       config=test_config)
-        amazon_session.login()
-        amazon_orders = AmazonOrders(amazon_session)
-        amazon_transactions = AmazonTransactions(amazon_session)
-
-        # WHEN
-        with open(test_config.cookie_jar_path, "r") as f:
-            cookies = json.loads(f.read())
-        cookies["session-token"] = "invalid-token"
-        cookies["x-main"] = "invalid-token"
-        with open(test_config.cookie_jar_path, "w") as f:
-            f.write(json.dumps(cookies))
-        amazon_session.session.cookies.update(cookies)
-
-        # THEN
-        with self.assertRaises(AmazonOrdersAuthError):
-            amazon_orders.get_order(order_id="1234-fake-id")
-        with self.assertRaises(AmazonOrdersAuthError):
-            amazon_orders.get_order_history(year=self.year,
-                                            start_index=self.start_index)
-        with self.assertRaises(AmazonOrdersAuthError):
-            amazon_transactions.get_transactions(days=self.transactions_days)
 
     def test_get_order_history(self):
         # WHEN