+++ tests/test_session.py
@@ -94,7 +94,6 @@ class TestSession(UnitTestCase):
     @patch("builtins.input")
     def test_mfa(self, input_mock):
         # GIVEN
-        os.environ.pop("AMAZON_OTP_SECRET_KEY", None)
         with open(os.path.join(self.RESOURCES_DIR, "auth", "signin.html"), "r", encoding="utf-8") as f:
             resp1 = responses.add(
                 responses.GET,
@@ -131,7 +130,6 @@ class TestSession(UnitTestCase):
     @patch("builtins.input")
     def test_new_otp(self, input_mock):
         # GIVEN
-        os.environ.pop("AMAZON_OTP_SECRET_KEY", None)
         with open(os.path.join(self.RESOURCES_DIR, "auth", "signin.html"), "r", encoding="utf-8") as f:
             resp1 = responses.add(
                 responses.GET,
@@ -349,7 +347,6 @@ class TestSession(UnitTestCase):
     @patch("builtins.input")
     def test_captcha_otp(self, input_mock):
         # GIVEN
-        os.environ.pop("AMAZON_OTP_SECRET_KEY", None)
         with open(os.path.join(self.RESOURCES_DIR, "auth", "signin.html"), "r", encoding="utf-8") as f:
             resp1 = responses.add(
                 responses.GET,