+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/4.0.0...HEAD)
 
 ### Added
-- Stability improvements.
+- Stability improvements around cookies, authentication, and logout.
 
 ## [4.0.0](https://github.com/alexdlaird/amazon-orders/compare/3.2.19...4.0.0) - 2025-04-21